video erweitert

+++ src/main/java/com/example/backend/repository/VideoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.example.backend.entity.Video;
 
 @Repository
-public interface VideoRepository extends CrudRepository<Video, Integer>{
+public interface VideoRepository extends CrudRepository<Video, Long>{
 
 }
